fix: standardize integer types in migrations to resolve join type mismatches
Standardized database column types across migrations to ensure type consistency
in joins and improve query performance. Changes include:

Primary Keys:
- Converted all auto-incrementing primary keys from bigint to integer for
  consistency and optimized storage
- Updated: global_compositions_id, global_hero_draft_order_id,
  composition_id, global_hero_talents_versus_heroes_id,
  global_hero_talents_with_heroes_id, global_hero_matchups_enemy_id,
  global_hero_matchups_ally_id, talent_id

Foreign Keys & Reference Columns:
- Changed heroes.id from integer to tinyInteger (supports 0-255 heroes)
- Changed maps.map_id from bigint to tinyInteger (supports 0-255 maps)
- Updated all hero reference columns to tinyInteger for consistency:
  * global_compositions.hero
  * global_hero_draft_order.hero
  * global_hero_talents_versus_heroes.enemy
  * global_hero_talents_with_heroes.ally

Composition Roles:
- Changed compositions.role_* columns from unsignedBigInteger to integer
  to better match mmr_type_ids.mmr_type_id

These changes resolve type mismatches in:
- GlobalCompositionsController
- GlobalDraftController
- GlobalHeroMapStatsController
- GlobalHeroMatchupsTalentsController

Benefits:
- Eliminates implicit type conversions in joins
- Reduces index and table sizes
- Improves query performance
- Ensures data consistency

+++ database/migrations/2023_12_01_200818_create_global_compositions_table.php
@@ -23,7 +23,7 @@ class CreateGlobalCompositionsTable extends Migration
             $table->tinyInteger('game_map');
             $table->unsignedInteger('hero_level');
             $table->integer('composition_id');
-            $table->integer('hero')->nullable();
+            $table->tinyInteger('hero')->nullable();
             $table->tinyInteger('mirror')->default(0);
             $table->integer('region');
             $table->tinyInteger('win_loss');
